@@ -1,6 +1,6 @@
-const Cookie = require('tough-cookie').Cookie;
+import { Cookie } from 'tough-cookie';
 
-function formatCookie(cookie) {
+export function formatCookie(cookie) {
   let expiresISO;
 
   if (cookie.expires instanceof Date) {
@@ -23,7 +23,7 @@ function formatCookie(cookie) {
     domain: cookie.domain || undefined, // must be undefined, not null, to exclude empty domain
     expires: expiresISO,
     httpOnly: cookie.httpOnly,
-    secure: cookie.secure
+    secure: cookie.secure,
   };
 }
 
@@ -37,19 +37,13 @@ function parseCookie(cookieString) {
 }
 
 function splitAndParse(header, divider) {
-  return header
-    .split(divider)
-    .filter(Boolean)
-    .map(parseCookie)
-    .filter(Boolean);
+  return header.split(divider).filter(Boolean).map(parseCookie).filter(Boolean);
 }
 
-module.exports = {
-  parseRequestCookies(cookieHeader) {
-    return splitAndParse(cookieHeader, ';');
-  },
-  parseResponseCookies(cookieHeader) {
-    return splitAndParse(cookieHeader, '\n');
-  },
-  formatCookie
-};
+export function parseRequestCookies(cookieHeader) {
+  return splitAndParse(cookieHeader, ';');
+}
+
+export function parseResponseCookies(cookieHeader) {
+  return splitAndParse(cookieHeader, '\n');
+}
