@@ -1,15 +1,14 @@
-'use strict';
 const max = Math.max;
 
-const {
+import {
   calculateRequestHeaderSize,
   calculateResponseHeaderSize,
   getHeaderValue,
-  parseHeaders
-} = require('./headers');
+  parseHeaders,
+} from './headers.js';
 
-const { parseRequestCookies, parseResponseCookies } = require('./cookies');
-const { isHttp1x, formatMillis } = require('./util');
+import { parseRequestCookies, parseResponseCookies } from './cookies.js';
+import { isHttp1x, formatMillis } from './util.js';
 function firstNonNegative(values) {
   for (let i = 0; i < values.length; ++i) {
     if (values[i] >= 0) return values[i];
@@ -25,7 +24,7 @@ function formatIP(ipAddress) {
   return ipAddress.replace(/^\[|]$/g, '');
 }
 
-module.exports = function(entry, response, page, options) {
+export default function (entry, response, page, options) {
   const responseHeaders = response.headers;
   const cookieHeader = getHeaderValue(responseHeaders, 'Set-Cookie');
 
@@ -36,7 +35,7 @@ module.exports = function(entry, response, page, options) {
     if (entry.extraResponseInfo.blockedCookies) {
       cookies = cookies.filter(
         ({ name }) =>
-          !entry.extraResponseInfo.blockedCookies.find(blockedCookie => {
+          !entry.extraResponseInfo.blockedCookies.find((blockedCookie) => {
             if (blockedCookie.cookie) {
               return blockedCookie.cookie.name === name;
             } else if (blockedCookie.cookieLine) {
@@ -47,7 +46,7 @@ module.exports = function(entry, response, page, options) {
             }
 
             return false;
-          })
+          }),
       );
     }
 
@@ -73,7 +72,7 @@ module.exports = function(entry, response, page, options) {
       encoding: response.encoding,
       mimeType: response.mimeType,
       size: text != null ? text.length : 0,
-      text: text
+      text: text,
     },
     headersSize: -1,
     bodySize: -1,
@@ -83,7 +82,7 @@ module.exports = function(entry, response, page, options) {
     fromDiskCache: response.fromDiskCache || false,
     fromEarlyHints: response.fromEarlyHints || false,
     fromServiceWorker: response.fromServiceWorker || false,
-    fromPrefetchCache: response.fromPrefetchCache || false
+    fromPrefetchCache: response.fromPrefetchCache || false,
   };
 
   const locationHeaderValue = getHeaderValue(responseHeaders, 'Location');
@@ -104,7 +103,7 @@ module.exports = function(entry, response, page, options) {
       entry.cache.beforeRequest = {
         lastAccess: '',
         eTag: '',
-        hitCount: 0
+        hitCount: 0,
       };
     }
   } else {
@@ -147,7 +146,11 @@ module.exports = function(entry, response, page, options) {
   const timing = response.timing;
   if (timing) {
     const blocked = formatMillis(
-      firstNonNegative([timing.dnsStart, timing.connectStart, timing.sendStart])
+      firstNonNegative([
+        timing.dnsStart,
+        timing.connectStart,
+        timing.sendStart,
+      ]),
     );
 
     const dns = parseOptionalTime(timing, 'dnsStart', 'dnsEnd');
@@ -165,7 +168,7 @@ module.exports = function(entry, response, page, options) {
       send,
       wait,
       receive,
-      ssl
+      ssl,
     };
 
     entry._requestTime = timing.requestTime;
@@ -207,8 +210,8 @@ module.exports = function(entry, response, page, options) {
       wait: 0,
       receive: 0,
       ssl: -1,
-      comment: 'No timings available from Chrome'
+      comment: 'No timings available from Chrome',
     };
     entry.time = 0;
   }
-};
+}
