@@ -1,11 +1,11 @@
-const { isHttp1x, formatMillis } = require('./util');
+import { isHttp1x, formatMillis } from './util.js';
 
 const max = Math.max;
 
-module.exports = function(entry, params) {
+export default function (entry, params) {
   const timings = entry.timings || {};
   timings.receive = formatMillis(
-    (params.timestamp - entry._requestTime) * 1000 - entry.__receiveHeadersEnd
+    (params.timestamp - entry._requestTime) * 1000 - entry.__receiveHeadersEnd,
   );
   entry.time =
     max(0, timings.blocked) +
@@ -31,11 +31,11 @@ module.exports = function(entry, params) {
 
       const compression = Math.max(
         0,
-        response.content.size - response.bodySize
+        response.content.size - response.bodySize,
       );
       if (compression > 0) {
         response.content.compression = compression;
       }
     }
   }
-};
+}
