@@ -1,9 +1,9 @@
 'use strict';
 
-const { name } = require('../package');
-const debug = require('debug')(name);
+import debug from 'debug';
+const log = debug('chrome-har');
 
-module.exports = function(method) {
+export default function (method) {
   switch (method) {
     case 'Network.webSocketCreated':
     case 'Network.webSocketFrameSent':
@@ -41,7 +41,7 @@ module.exports = function(method) {
       // ignore
       break;
     default:
-      debug(`Unhandled event: ${method}`);
+      log(`Unhandled event: ${method}`);
       break;
   }
-};
+}
